Combined words listed multiple times with different English Definitions together, and logic needed to make this work.
    get_pop_words.py
Added logic to check if a word is already been seen, and add a translation if it has.
Added logic to keep words in order using heap when creating output data.

    csv files
Updated with changes generated from script due to combining repeated words.

    italianTopWordsGame.py & russianTopWordsGame.py
Updated logic to work with new csv files.
Added logic to show all possible answers if correct, and all answers if wrong.
Added italics to word to guess.